refactor: split db.rs into per-domain submodules
db.rs had grown to ~2100 lines covering schema/migrations,
validation, and CRUD for four unrelated tables in one file. Split
into db/{schema,validation,transcriptions,dictionary,snippets,
cleanup_cache}.rs with mod.rs re-exporting everything so the
db::* call sites used throughout the codebase are unchanged.

Also changed open() to accept impl AsRef<Path> instead of &str,
removing a needless to_str().unwrap() at every call site.
