+++ src-tauri/src/data/db/schema.rs
@@ -1,0 +1,448 @@
+//! Database schema definition, connection `open`, and versioned migrations.
+
+use anyhow::Result;
+use rusqlite::{params, Connection};
+use std::sync::{Arc, Mutex};
+
+use super::*;
+
+const SCHEMA: &str = "
+CREATE TABLE IF NOT EXISTS transcriptions (
+  id          INTEGER PRIMARY KEY AUTOINCREMENT,
+  raw_text    TEXT    NOT NULL,
+  clean_text  TEXT    NOT NULL,
+  words       INTEGER NOT NULL DEFAULT 0,
+  spoken_words INTEGER,
+  duration_ms INTEGER NOT NULL DEFAULT 0,
+  api_used    TEXT    NOT NULL DEFAULT '',
+  created_at  DATETIME NOT NULL DEFAULT (datetime('now'))
+);
+CREATE INDEX IF NOT EXISTS idx_transcriptions_created_at
+  ON transcriptions(created_at);
+CREATE TABLE IF NOT EXISTS lifetime_stats (
+  id          INTEGER PRIMARY KEY CHECK (id = 1),
+  total_words INTEGER NOT NULL DEFAULT 0
+);
+CREATE TABLE IF NOT EXISTS dictionary (
+  id               INTEGER PRIMARY KEY AUTOINCREMENT,
+  term             TEXT    NOT NULL UNIQUE,
+  mistake          TEXT,
+  auto_learned     INTEGER NOT NULL DEFAULT 0,
+  correction_count INTEGER NOT NULL DEFAULT 0,
+  confidence_tier  TEXT    NOT NULL DEFAULT 'low',
+  last_seen_at     DATETIME,
+  created_at       DATETIME NOT NULL DEFAULT (datetime('now'))
+);
+CREATE TABLE IF NOT EXISTS snippets (
+  id           INTEGER PRIMARY KEY AUTOINCREMENT,
+  trigger      TEXT    NOT NULL UNIQUE,
+  expansion    TEXT    NOT NULL,
+  instructions TEXT    NOT NULL DEFAULT '',
+  use_count    INTEGER NOT NULL DEFAULT 0,
+  created_at   DATETIME NOT NULL DEFAULT (datetime('now'))
+);
+CREATE TABLE IF NOT EXISTS pending_corrections (
+  id           INTEGER PRIMARY KEY AUTOINCREMENT,
+  wrong_word   TEXT    NOT NULL,
+  correct_word TEXT    NOT NULL,
+  created_at   DATETIME NOT NULL DEFAULT (datetime('now'))
+);
+CREATE INDEX IF NOT EXISTS idx_pending_words
+  ON pending_corrections(wrong_word, correct_word);
+CREATE TABLE IF NOT EXISTS auto_learn_events (
+  id             INTEGER PRIMARY KEY AUTOINCREMENT,
+  event_type     TEXT    NOT NULL,
+  reason_code    TEXT    NOT NULL DEFAULT '',
+  app_context    TEXT    NOT NULL DEFAULT '',
+  mistake_hash   TEXT    NOT NULL DEFAULT '',
+  correction_hash TEXT   NOT NULL DEFAULT '',
+  confidence     REAL    NOT NULL DEFAULT 0.0,
+  created_at     DATETIME NOT NULL DEFAULT (datetime('now'))
+);
+CREATE INDEX IF NOT EXISTS idx_auto_learn_events_event_type
+  ON auto_learn_events(event_type, created_at);
+CREATE TABLE IF NOT EXISTS auto_learn_candidates (
+  id                 INTEGER PRIMARY KEY AUTOINCREMENT,
+  wrong_word         TEXT    NOT NULL,
+  correct_word       TEXT    NOT NULL,
+  confidence_sum     REAL    NOT NULL DEFAULT 0.0,
+  confidence_avg     REAL    NOT NULL DEFAULT 0.0,
+  seen_count         INTEGER NOT NULL DEFAULT 0,
+  last_seen_at       DATETIME NOT NULL DEFAULT (datetime('now')),
+  cooldown_until     DATETIME,
+  promoted_at        DATETIME,
+  UNIQUE(wrong_word, correct_word)
+);
+CREATE INDEX IF NOT EXISTS idx_auto_learn_candidates_seen
+  ON auto_learn_candidates(last_seen_at);
+CREATE TABLE IF NOT EXISTS cleanup_cache (
+  key         TEXT PRIMARY KEY,
+  clean_text  TEXT NOT NULL,
+  hit_count   INTEGER NOT NULL DEFAULT 0,
+  created_at  DATETIME NOT NULL DEFAULT (datetime('now')),
+  last_hit_at DATETIME NOT NULL DEFAULT (datetime('now')),
+  expires_at  DATETIME NOT NULL,
+  created_at_epoch  INTEGER,
+  last_hit_at_epoch INTEGER,
+  expires_at_epoch  INTEGER,
+  is_snippet  INTEGER NOT NULL DEFAULT 0
+);
+CREATE INDEX IF NOT EXISTS idx_cleanup_cache_expires_at
+  ON cleanup_cache(expires_at);
+CREATE INDEX IF NOT EXISTS idx_cleanup_cache_last_hit_at
+  ON cleanup_cache(last_hit_at);
+";
+
+pub fn open(path: impl AsRef<std::path::Path>) -> Result<Db> {
+    let db_path = path.as_ref();
+    let mut conn = Connection::open(db_path)?;
+    conn.execute_batch(SCHEMA)?;
+    conn.execute_batch("PRAGMA journal_mode=WAL;")?;
+
+    let user_version: i64 = conn.query_row("PRAGMA user_version", [], |r| r.get(0))?;
+
+    if user_version < 2 {
+        if db_path.exists() {
+            let _ = std::fs::copy(db_path, db_path.with_extension("db.bak"));
+        }
+
+        // IMPORTANT: each migration uses BEGIN/COMMIT for atomicity, followed by an
+        // explicit ROLLBACK. If the migration fails mid-way, sqlite3_exec aborts but
+        // leaves the BEGIN open. Without the ROLLBACK cleanup, every subsequent INSERT
+        // would execute inside that ghost transaction and be silently discarded on
+        // connection close — causing all user data to vanish on restart.
+        let _ = conn.execute_batch(
+            "ALTER TABLE snippets ADD COLUMN instructions TEXT NOT NULL DEFAULT '';",
+        );
+        let _ = conn.execute_batch(
+            "CREATE TABLE IF NOT EXISTS pending_corrections (
+               id           INTEGER PRIMARY KEY AUTOINCREMENT,
+               wrong_word   TEXT    NOT NULL,
+               correct_word TEXT    NOT NULL,
+               created_at   DATETIME NOT NULL DEFAULT (datetime('now'))
+             );
+             CREATE INDEX IF NOT EXISTS idx_pending_words
+               ON pending_corrections(wrong_word, correct_word);",
+        );
+        // Migrate dictionary to final schema: term (required) + mistake (optional).
+        // Handles all prior states (original wrong/correct columns, or already migrated).
+        let _ = conn.execute_batch(
+            "BEGIN;
+             CREATE TABLE IF NOT EXISTS dictionary_v3 (
+                 id               INTEGER PRIMARY KEY AUTOINCREMENT,
+                 term             TEXT    NOT NULL UNIQUE,
+                 mistake          TEXT,
+                 auto_learned     INTEGER NOT NULL DEFAULT 0,
+                 correction_count INTEGER NOT NULL DEFAULT 0,
+                 created_at       DATETIME NOT NULL DEFAULT (datetime('now'))
+             );
+             INSERT OR IGNORE INTO dictionary_v3
+                 (id, term, mistake, auto_learned, correction_count, created_at)
+                 SELECT id,
+                        COALESCE(correct, wrong),
+                        CASE WHEN correct IS NOT NULL THEN wrong ELSE NULL END,
+                        auto_learned, correction_count, created_at
+                 FROM dictionary;
+             DROP TABLE dictionary;
+             ALTER TABLE dictionary_v3 RENAME TO dictionary;
+             COMMIT;",
+        );
+        // Clean up any dangling transaction left by a failed migration above.
+        let _ = conn.execute_batch("ROLLBACK;");
+        conn.execute_batch("PRAGMA user_version = 2;")?;
+    }
+    if user_version < 3 {
+        let res = conn.execute_batch(
+            "BEGIN;
+             CREATE TABLE IF NOT EXISTS cleanup_cache (
+               key         TEXT PRIMARY KEY,
+               clean_text  TEXT NOT NULL,
+               hit_count   INTEGER NOT NULL DEFAULT 0,
+               created_at  DATETIME NOT NULL DEFAULT (datetime('now')),
+               last_hit_at DATETIME NOT NULL DEFAULT (datetime('now')),
+               expires_at  DATETIME NOT NULL,
+               created_at_epoch  INTEGER,
+               last_hit_at_epoch INTEGER,
+               expires_at_epoch  INTEGER
+              );
+             CREATE INDEX IF NOT EXISTS idx_cleanup_cache_expires_at
+               ON cleanup_cache(expires_at);
+             CREATE INDEX IF NOT EXISTS idx_cleanup_cache_last_hit_at
+               ON cleanup_cache(last_hit_at);
+             PRAGMA user_version = 3;
+             COMMIT;",
+        );
+        if let Err(err) = res {
+            let _ = conn.execute_batch("ROLLBACK;");
+            return Err(err.into());
+        }
+    }
+    if user_version < 4 {
+        conn.execute_batch("BEGIN;")?;
+        if let Err(err) = (|| -> Result<()> {
+            ensure_table_column(
+                &conn,
+                "dictionary",
+                "confidence_tier",
+                "ALTER TABLE dictionary ADD COLUMN confidence_tier TEXT NOT NULL DEFAULT 'low';",
+            )?;
+            ensure_table_column(
+                &conn,
+                "dictionary",
+                "last_seen_at",
+                "ALTER TABLE dictionary ADD COLUMN last_seen_at DATETIME;",
+            )?;
+            conn.execute_batch(
+                "CREATE TABLE IF NOT EXISTS auto_learn_events (
+                   id             INTEGER PRIMARY KEY AUTOINCREMENT,
+                   event_type     TEXT    NOT NULL,
+                   reason_code    TEXT    NOT NULL DEFAULT '',
+                   app_context    TEXT    NOT NULL DEFAULT '',
+                   mistake_hash   TEXT    NOT NULL DEFAULT '',
+                   correction_hash TEXT   NOT NULL DEFAULT '',
+                   confidence     REAL    NOT NULL DEFAULT 0.0,
+                   created_at     DATETIME NOT NULL DEFAULT (datetime('now'))
+                 );
+                 CREATE INDEX IF NOT EXISTS idx_auto_learn_events_event_type
+                   ON auto_learn_events(event_type, created_at);
+                 CREATE TABLE IF NOT EXISTS auto_learn_candidates (
+                   id                 INTEGER PRIMARY KEY AUTOINCREMENT,
+                   wrong_word         TEXT    NOT NULL,
+                   correct_word       TEXT    NOT NULL,
+                   confidence_sum     REAL    NOT NULL DEFAULT 0.0,
+                   confidence_avg     REAL    NOT NULL DEFAULT 0.0,
+                   seen_count         INTEGER NOT NULL DEFAULT 0,
+                   last_seen_at       DATETIME NOT NULL DEFAULT (datetime('now')),
+                   cooldown_until     DATETIME,
+                   promoted_at        DATETIME,
+                   UNIQUE(wrong_word, correct_word)
+                 );
+                 CREATE INDEX IF NOT EXISTS idx_auto_learn_candidates_seen
+                   ON auto_learn_candidates(last_seen_at);
+                 PRAGMA user_version = 4;",
+            )?;
+            Ok(())
+        })() {
+            let _ = conn.execute_batch("ROLLBACK;");
+            return Err(err);
+        }
+        conn.execute_batch("COMMIT;")?;
+    }
+    if user_version < 5 {
+        conn.execute_batch("BEGIN;")?;
+        if let Err(err) = (|| -> Result<()> {
+            ensure_cleanup_cache_schema(&conn)?;
+            conn.execute_batch("PRAGMA user_version = 5;")?;
+            Ok(())
+        })() {
+            let _ = conn.execute_batch("ROLLBACK;");
+            return Err(err);
+        }
+        conn.execute_batch("COMMIT;")?;
+    }
+    if user_version < 6 {
+        conn.execute_batch("BEGIN;")?;
+        if let Err(err) = (|| -> Result<()> {
+            ensure_cleanup_cache_schema(&conn)?;
+            conn.execute_batch("PRAGMA user_version = 6;")?;
+            Ok(())
+        })() {
+            let _ = conn.execute_batch("ROLLBACK;");
+            return Err(err);
+        }
+        conn.execute_batch("COMMIT;")?;
+    }
+    if user_version < 7 {
+        conn.execute_batch("BEGIN;")?;
+        if let Err(err) = (|| -> Result<()> {
+            ensure_table_column(
+                &conn,
+                "transcriptions",
+                "spoken_words",
+                "ALTER TABLE transcriptions ADD COLUMN spoken_words INTEGER;",
+            )?;
+            backfill_spoken_words(&conn)?;
+            conn.execute_batch("PRAGMA user_version = 7;")?;
+            Ok(())
+        })() {
+            let _ = conn.execute_batch("ROLLBACK;");
+            return Err(err);
+        }
+        conn.execute_batch("COMMIT;")?;
+    }
+    if user_version < 8 {
+        let res = conn.execute_batch(
+            "BEGIN;
+             CREATE INDEX IF NOT EXISTS idx_transcriptions_created_at
+               ON transcriptions(created_at);
+             PRAGMA user_version = 8;
+             COMMIT;",
+        );
+        if let Err(err) = res {
+            let _ = conn.execute_batch("ROLLBACK;");
+            return Err(err.into());
+        }
+    }
+    if user_version < 9 {
+        // Seed the lifetime word counter from existing history so upgrading
+        // users don't see it reset to zero. From here on it's only ever
+        // incremented on insert, never recomputed from transcriptions, so
+        // history retention pruning can't shrink it.
+        let res = conn.execute_batch(
+            "BEGIN;
+             INSERT OR IGNORE INTO lifetime_stats (id, total_words)
+               SELECT 1, COALESCE(SUM(words), 0) FROM transcriptions;
+             PRAGMA user_version = 9;
+             COMMIT;",
+        );
+        if let Err(err) = res {
+            let _ = conn.execute_batch("ROLLBACK;");
+            return Err(err.into());
+        }
+    }
+    ensure_cleanup_cache_schema(&conn)?;
+
+    // Self-heal: some databases ended up with user_version >= 7 without the
+    // spoken_words column from that migration actually landing (an
+    // interrupted migration during the Verenu rename/update). The column
+    // addition and backfill run in one transaction so a failed backfill
+    // rolls back the column too, letting this retry on the next launch.
+    // Errors are propagated: without this column, every transcription
+    // insert fails, so running in this state is worse than failing to open.
+    if !table_has_column(&conn, "transcriptions", "spoken_words")? {
+        let tx = conn.transaction()?;
+        let res = (|| -> Result<()> {
+            tx.execute_batch("ALTER TABLE transcriptions ADD COLUMN spoken_words INTEGER;")?;
+            backfill_spoken_words(&tx)?;
+            Ok(())
+        })();
+        match res {
+            Ok(()) => {
+                tx.commit()?;
+            }
+            Err(err) => {
+                log::error!("Failed to self-heal spoken_words column: {err}");
+                return Err(err);
+            }
+        }
+    }
+
+    Ok(Arc::new(Mutex::new(conn)))
+}
+
+pub fn table_has_column(conn: &Connection, table: &str, column: &str) -> Result<bool> {
+    let pragma = format!("PRAGMA table_info({table})");
+    let mut stmt = conn.prepare(&pragma)?;
+    let mut rows = stmt.query([])?;
+    while let Some(row) = rows.next()? {
+        let name: String = row.get(1)?;
+        if name == column {
+            return Ok(true);
+        }
+    }
+    Ok(false)
+}
+
+fn ensure_table_column(
+    conn: &Connection,
+    table: &str,
+    column: &str,
+    def_sql: &str,
+) -> Result<bool> {
+    if table_has_column(conn, table, column)? {
+        return Ok(false);
+    }
+    conn.execute_batch(def_sql)?;
+    Ok(true)
+}
+
+fn ensure_cleanup_cache_schema(conn: &Connection) -> Result<()> {
+    let mut repaired = false;
+    repaired |= ensure_table_column(
+        conn,
+        "cleanup_cache",
+        "created_at_epoch",
+        "ALTER TABLE cleanup_cache ADD COLUMN created_at_epoch INTEGER;",
+    )?;
+    repaired |= ensure_table_column(
+        conn,
+        "cleanup_cache",
+        "last_hit_at_epoch",
+        "ALTER TABLE cleanup_cache ADD COLUMN last_hit_at_epoch INTEGER;",
+    )?;
+    repaired |= ensure_table_column(
+        conn,
+        "cleanup_cache",
+        "expires_at_epoch",
+        "ALTER TABLE cleanup_cache ADD COLUMN expires_at_epoch INTEGER;",
+    )?;
+    repaired |= ensure_table_column(
+        conn,
+        "cleanup_cache",
+        "is_snippet",
+        "ALTER TABLE cleanup_cache ADD COLUMN is_snippet INTEGER NOT NULL DEFAULT 0;",
+    )?;
+    if repaired {
+        conn.execute_batch(
+            "UPDATE cleanup_cache
+             SET created_at_epoch = COALESCE(created_at_epoch, CAST(strftime('%s', created_at || 'Z') AS INTEGER)),
+                 last_hit_at_epoch = COALESCE(last_hit_at_epoch, CAST(strftime('%s', last_hit_at || 'Z') AS INTEGER)),
+                 expires_at_epoch = COALESCE(expires_at_epoch, CAST(strftime('%s', expires_at || 'Z') AS INTEGER))
+             WHERE created_at_epoch IS NULL
+                OR last_hit_at_epoch IS NULL
+                OR expires_at_epoch IS NULL;",
+        )?;
+    }
+    conn.execute_batch(
+        "CREATE INDEX IF NOT EXISTS idx_cleanup_cache_expires_at_epoch
+           ON cleanup_cache(expires_at_epoch);
+         CREATE INDEX IF NOT EXISTS idx_cleanup_cache_last_hit_at_epoch
+           ON cleanup_cache(last_hit_at_epoch);",
+    )?;
+    Ok(())
+}
+
+fn load_snippet_rows(conn: &Connection) -> Result<Vec<Snippet>> {
+    let mut snippet_stmt = conn.prepare(
+        "SELECT id, trigger, expansion, instructions, use_count, created_at \
+         FROM snippets",
+    )?;
+    let rows = snippet_stmt
+        .query_map([], |r| {
+            Ok(Snippet {
+                id: r.get(0)?,
+                trigger: r.get(1)?,
+                expansion: r.get(2)?,
+                instructions: r.get(3)?,
+                use_count: r.get(4)?,
+                created_at: r.get(5)?,
+            })
+        })?
+        .collect::<rusqlite::Result<Vec<_>>>()?;
+    Ok(rows)
+}
+
+pub fn compute_spoken_words(conn: &Connection, raw_text: &str) -> Result<i64> {
+    let snippets = load_snippet_rows(conn)?;
+    Ok(crate::data::snippets::count_words_without_snippet_triggers(
+        raw_text, &snippets,
+    ))
+}
+
+fn backfill_spoken_words(conn: &Connection) -> Result<()> {
+    let snippets = load_snippet_rows(conn)?;
+    let mut select =
+        conn.prepare("SELECT id, raw_text FROM transcriptions WHERE spoken_words IS NULL")?;
+    let rows = select
+        .query_map([], |r| Ok((r.get::<_, i64>(0)?, r.get::<_, String>(1)?)))?
+        .collect::<rusqlite::Result<Vec<_>>>()?;
+    let mut update = conn.prepare("UPDATE transcriptions SET spoken_words = ?2 WHERE id = ?1")?;
+
+    for (id, raw_text) in rows {
+        let spoken_words =
+            crate::data::snippets::count_words_without_snippet_triggers(&raw_text, &snippets);
+        update.execute(params![id, spoken_words])?;
+    }
+
+    Ok(())
+}
