+++ src-tauri/src/data/db/snippets.rs
@@ -1,0 +1,158 @@
+//! Snippet CRUD and use-count tracking.
+
+use anyhow::Result;
+use rusqlite::params;
+use serde::{Deserialize, Serialize};
+
+use super::*;
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct Snippet {
+    pub id: i64,
+    pub trigger: String,
+    pub expansion: String,
+    pub instructions: String,
+    pub use_count: i64,
+    pub created_at: String,
+}
+
+#[cfg(test)]
+pub fn insert_snippet(db: &Db, trigger: &str, expansion: &str, instructions: &str) -> Result<()> {
+    let normalized_trigger = require_nonempty_trimmed("Trigger", trigger)?;
+    validate_char_limit("Trigger", &normalized_trigger, SNIPPET_TRIGGER_CHAR_LIMIT)?;
+    let normalized_expansion = normalize_multiline(expansion);
+    if normalized_expansion.is_empty() {
+        return Err(anyhow::anyhow!("Expansion cannot be empty"));
+    }
+    let normalized_instructions = normalize_multiline(instructions);
+
+    let conn = lock_conn(db)?;
+    conn.execute(
+        "INSERT INTO snippets (trigger, expansion, instructions) VALUES (?1, ?2, ?3)",
+        params![
+            normalized_trigger,
+            normalized_expansion,
+            normalized_instructions
+        ],
+    )?;
+    Ok(())
+}
+
+pub fn insert_snippet_returning(
+    db: &Db,
+    trigger: &str,
+    expansion: &str,
+    instructions: &str,
+) -> Result<CreatedRecordMeta> {
+    let normalized_trigger = require_nonempty_trimmed("Trigger", trigger)?;
+    validate_char_limit("Trigger", &normalized_trigger, SNIPPET_TRIGGER_CHAR_LIMIT)?;
+    let normalized_expansion = normalize_multiline(expansion);
+    if normalized_expansion.is_empty() {
+        return Err(anyhow::anyhow!("Expansion cannot be empty"));
+    }
+    let normalized_instructions = normalize_multiline(instructions);
+
+    // Insert and read last_insert_rowid under a single lock to prevent another
+    // thread's insert racing between the two acquisitions and returning the wrong id.
+    let conn = lock_conn(db)?;
+    conn.execute(
+        "INSERT INTO snippets (trigger, expansion, instructions) VALUES (?1, ?2, ?3)",
+        params![
+            normalized_trigger,
+            normalized_expansion,
+            normalized_instructions
+        ],
+    )?;
+    let id = conn.last_insert_rowid();
+    let created_at = conn.query_row(
+        "SELECT created_at FROM snippets WHERE id=?1",
+        params![id],
+        |r| r.get(0),
+    )?;
+    Ok(CreatedRecordMeta { id, created_at })
+}
+
+pub fn update_snippet(
+    db: &Db,
+    id: i64,
+    trigger: &str,
+    expansion: &str,
+    instructions: &str,
+) -> Result<()> {
+    let normalized_trigger = require_nonempty_trimmed("Trigger", trigger)?;
+    validate_char_limit("Trigger", &normalized_trigger, SNIPPET_TRIGGER_CHAR_LIMIT)?;
+    let normalized_expansion = normalize_multiline(expansion);
+    if normalized_expansion.is_empty() {
+        return Err(anyhow::anyhow!("Expansion cannot be empty"));
+    }
+    let normalized_instructions = normalize_multiline(instructions);
+
+    let conn = lock_conn(db)?;
+    let changed = conn.execute(
+        "UPDATE snippets SET trigger=?2, expansion=?3, instructions=?4 WHERE id=?1",
+        params![
+            id,
+            normalized_trigger,
+            normalized_expansion,
+            normalized_instructions
+        ],
+    )?;
+    require_row_changed(changed, "Snippet", id)?;
+    Ok(())
+}
+
+pub fn delete_snippet(db: &Db, id: i64) -> Result<()> {
+    let conn = lock_conn(db)?;
+    let changed = conn.execute("DELETE FROM snippets WHERE id=?1", params![id])?;
+    require_row_changed(changed, "Snippet", id)?;
+    Ok(())
+}
+
+pub fn query_snippets(db: &Db) -> Result<Vec<Snippet>> {
+    let conn = lock_conn(db)?;
+    let mut stmt = conn.prepare(
+        "SELECT id, trigger, expansion, instructions, use_count, created_at \
+         FROM snippets ORDER BY created_at DESC",
+    )?;
+    let rows = stmt
+        .query_map([], |r| {
+            Ok(Snippet {
+                id: r.get(0)?,
+                trigger: r.get(1)?,
+                expansion: r.get(2)?,
+                instructions: r.get(3)?,
+                use_count: r.get(4)?,
+                created_at: r.get(5)?,
+            })
+        })?
+        .collect::<rusqlite::Result<Vec<_>>>()?;
+    Ok(rows)
+}
+
+pub fn increment_snippet_use(db: &Db, id: i64) -> Result<()> {
+    let conn = lock_conn(db)?;
+    conn.execute(
+        "UPDATE snippets SET use_count = use_count + 1 WHERE id=?1",
+        params![id],
+    )?;
+    Ok(())
+}
+
+pub fn increment_snippet_use_counts(db: &Db, counts: &[(i64, i64)]) -> Result<()> {
+    if counts.is_empty() {
+        return Ok(());
+    }
+    let mut conn = lock_conn(db)?;
+    let tx = conn.transaction()?;
+    {
+        let mut stmt = tx.prepare("UPDATE snippets SET use_count = use_count + ?2 WHERE id=?1")?;
+        for (id, count) in counts.iter().copied() {
+            if count <= 0 {
+                continue;
+            }
+            stmt.execute(params![id, count])?;
+        }
+    }
+    tx.commit()?;
+    Ok(())
+}
